fix(api): extract storage_report module + mirror empty StorageEntry on missing home (refs #2054)
Round-1 fix for Codex feedback on PR #2054.

P1 (module boundary): extracted the canonical storage scanner —
``HOME_SUBDIRS``, ``DirScan`` / ``HomeReport`` dataclasses, the
``scan_pollypm_home`` entry point, and every walk helper it depends
on — into a new neutral ``pollypm.storage_report`` module. Both
``pm storage report`` (CLI) and ``GET /api/v1/storage`` (web) now
consume the same source of truth without coupling the route to the
Typer / bootstrap-pg / migrate-to-pg / prune CLI surface.
``cli_features/storage.py`` re-imports the public names under the
original ``_``-prefixed identifiers so existing in-tree callers and
tests keep working.

P0 (subdir 404 on fresh install): the subdir endpoint previously
returned 404 with a misleading "bug in scan_pollypm_home" hint when
``~/.pollypm/`` didn't exist (fresh install / alternate
``base_dir``), while the top-level endpoint correctly returned 200
with empty totals. The subdir endpoint now mirrors that invariant:
return 200 with a zeroed ``StorageEntry`` for a valid canonical
subdir name when the scan produces no row for it.

Tests:
- new ``test_storage_subdir_returns_empty_when_home_missing`` covers
  every canonical subdir against a wiped home.
- new ``test_storage_route_does_not_import_cli_features`` greps the
  route source to lock in the module boundary.
- existing cap-hit tests patch the canonical
  ``storage_report.SCAN_FILE_CAP`` (the CLI re-export is a static
  bind and patching it has no effect on the scanner).
